requirements.txt
scrapy
beautifulsoup4
mysql-connector-python
flask